requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
scipy
umap-learn
hdbscan

# pip install pyclustertend # - please run separately.
# pip install category-encoders